% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_micro.R
\name{create_micro}
\alias{create_micro}
\title{Create a microclimate config file}
\usage{
create_micro(
  latitude,
  longitude,
  altitude_m,
  dates,
  daily_temp_max_c,
  daily_temp_min_c,
  daily_rh_max_perc,
  daily_rh_min_perc,
  daily_cloud_max_perc,
  daily_cloud_min_perc,
  daily_wind_max_ms,
  daily_wind_min_ms,
  daily_rainfall_mm,
  weather_height_m = 2,
  adult_height_m = 1,
  shade_prop = 0,
  even_rain = TRUE,
  soil_reflectance = 0.15,
  solar_attenuation_data = solar_attenuation(latitude, longitude)
)
}
\arguments{
\item{latitude, longitude, altitude_m}{scalars for the coordinates and altitude
at these coordinates, to compute solar radiation and air pressure}

\item{dates}{vector of dates corresponding to data vectors - these must be
the dates in Greenwich Mean Time (so that longitude can be used to compute
solar noon)}

\item{daily_temp_max_c}{vector of daily maximum temperature in centigrade}

\item{daily_temp_min_c}{vector of daily minimum temperature in centigrade}

\item{daily_rh_max_perc}{vector of daily maximum relative humidity in percent}

\item{daily_rh_min_perc}{vector of daily minimum relative humidity in percent}

\item{daily_cloud_max_perc}{vector of daily maximum cloud cover in percent}

\item{daily_cloud_min_perc}{vector of daily minimum cloud cover in percent}

\item{daily_wind_max_ms}{vector of daily maximum wind speed in m/s}

\item{daily_wind_min_ms}{vector of daily minimum wind speed in m/s}

\item{daily_rainfall_mm}{vector of daily rainfall in millimetres}

\item{weather_height_m}{scalar height in metres at which the weather data
were recorded}

\item{adult_height_m}{scalar height above ground experienced by adult
mosquitoes in metres}

\item{shade_prop}{proportion of full shade experienced by mosquitoes}

\item{even_rain}{whether to spread daily rainfall out evenly over 24. If
FALSE, have it all fall at midnight.}

\item{soil_reflectance}{soil solar reflectance, decimal \%}

\item{solar_attenuation_data}{dataframe giving on solar attenuation at the
111 wavelengths for the given location, as created using
\code{solar_attentuation()}}
}
\description{
Create a 'micro' file for input to NicheMapR::microclimate(), given vectors of
daily macroclimate data
}
